backend/requirements.txt
Flask==3.0.2
flask-sqlalchemy==3.1.1
numpy==1.24.4
pandas==2.0.3
yfinance==0.2.36
flask-apscheduler==1.13.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
markdown==3.5.2
flask-cors==4.0.0
python-dotenv==1.0.0
flask-session==0.7.0
redis==5.0.3
